+++ packages/page-builder-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # react-page-and-form-builder
 
+## 0.1.3
+
+### Patch Changes
+
+- enhance component integration and performance optimizations across PageBuilder and NodeRenderer
+
+## 0.1.2
+
+### Patch Changes
+
+- 7c9cd39: publish file
+# react-page-and-form-builder
+
 ## 0.1.2
 
 ### Patch Changes